server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.35.0
python-dotenv==1.0.0
yt-dlp==2023.12.30
gunicorn==21.2.0
requests==2.31.0
urllib3==2.0.7